refactor: move getPostFieldAddMessage to models module
