+++ src/commands/field-add-text.ts
@@ -4,9 +4,13 @@ import { capitalCase } from "change-case";
 
 import { getHost, getToken } from "../auth";
 import { CommandError, createCommand, type CommandConfig } from "../lib/command";
-import { resolveFieldTarget, resolveModel, TARGET_OPTIONS } from "../models";
+import {
+	getPostFieldAddMessage,
+	resolveFieldTarget,
+	resolveModel,
+	TARGET_OPTIONS,
+} from "../models";
 import { getRepositoryName } from "../project";
-import { getPostFieldAddMessage } from "./field-add";
 
 const config = {
 	name: "prismic field add text",
