+++ src/models.ts
@@ -435,3 +435,16 @@ function validateLeafField(id: string, fields: Record<string, Field>, context: s
 		);
 	}
 }
+
+export function getPostFieldAddMessage(args: {
+	targetId: string;
+	modelKind: "slice" | "customType";
+}): string {
+	const { modelKind, targetId } = args;
+
+	if (modelKind === "slice") {
+		return `Run \`prismic slice view ${targetId}\` to view the updated slice.`;
+	} else {
+		return `Run \`prismic type view ${targetId}\` to view the updated type.`;
+	}
+}
